requirements.txt
Flask==1.1.2
google-cloud-datastore==1.15.1
google-auth
requests_oauthlib
